revtools: Tools to support reviews and evidence synthesis

Researchers commonly need to summarize scientific information, a process known as 'evidence synthesis'. The first stage of a synthesis process (such as a systematic review or meta-analysis) is to download a list of references from academic search engines such as 'Web of Knowledge' or 'Scopus'. The traditional approach to systematic review is then to sort these data manually, first by locating and removing duplicated entries, and then screening to remove irrelevant content by viewing titles and abstracts (in that order). 'revtools' provides interfaces for each of these tasks. An alternative approach, however, is to draw on tools from machine learning to visualise patterns in the corpus. In this case, you can use 'revtools' to render ordinations of text drawn from article titles, keywords and abstracts, and interactively select or exclude individual references, words or topics.

Functions

Article screening

screen_duplicates Screen for duplicates

screen_titles Screen articles by title

screen_abstracts Screen articles by abstract

screen_topics Screen data by topic

Bibliographic methods

read_bibliography Import bibliographic data

write_bibliography Export bibliographic data

bibliography-class Format for storing bibliographic data

bibliography-methods Print, summary, as.bibliography, as.data.frame and [ methods for class 'bibliography'

tag_lookup Lookup table for ris tag transformations

merge_columns rbind two data.frames with different numbers of columns

format_citation Return a clean citation from a bibliography or data.frame

add_line_breaks Set a maximum character width for strings in a vector

Distributing tasks among a team

allocate_effort Specify how to distribute screening effort among a team of reviewers

distribute_tasks Split a dataset among a team of reviewers

aggregate_tasks Combine screening results from a team of reviewers

Duplicate detection

fuzz_functions Fuzzy string matching

find_duplicates Locate potentially duplicated references

extract_unique_references Return a data.frame with only 'unique' references

screen_duplicates Screen for duplicates

Text mining

revwords Stopwords used in revtools functions

make_dtm Construct a Document-Term Matrix from bibliographic data

run_topic_model Wrapper function for topic models

screen_topics Screen data by topic