Shiny app for locating and excluding duplicated entries in a dataset

This is a simple app for calculating, displaying and screening potential duplicates in bibliographic data

screen_duplicates(x, max_file_size)

Arguments

x

An (optional) object of class data.frame or bibliography to open in the browser. If empty, the app will launch with no data. Data can be added within the app via the 'import' button.

max_file_size

Optional argument to set the maximum file size (in MB) that the app will accept.

Value

This function launches a Shiny app in the users' default browser, allowing the user to customize their parameters for duplicate detection, and visualise the results.

Details

This app is effectively a wrapper for find_duplicates, with the added option to manually screen pairs of duplicates to check the results. Consequently, this is a more reliable method than extract_unique_references of dealing with the duplicates identified by find_duplicates, and for testing whether that function has returned sensible results for a given dataset.

See also

Examples

# to run the app and upload data interactively
if (FALSE) screen_duplicates()
# or to specify data from the workspace
file_location <- system.file(
  "extdata",
  "avian_ecology_bibliography.ris",
  package = "revtools")
x <- read_bibliography(file_location)
# to run the app using these data:
if (FALSE) screen_duplicates(x)
# or to run the app & save results to the workspace:
if (FALSE) result <- screen_duplicates(x)