Description of class 'screen_topics_progress'

screen_topics_progress is an S3 class designed to store data from screen_topics, allowing the user to re-load a previously calculated topic model. If you just want to save your decisions on article inclusion/exclusion, along with your notes, then this is probably overkill as that information can simply be exported as a .csv file.

slots

Class 'screen_topics_progress' has seven slots containing the following information:

raw duplicate of data passed to screen_topics

stopwords vector of words excluded from the dtm

columns vector of column names in the original dataset

grouped a data.frame showing grouped data as specified by the user

dtm document-term matrix, created by make_dtm

model most recent topic model, created by run_topic_model

plot_ready data needed for the main plot (coordinates etc.)