Lookup table for ris tags

ris-like bibliographic data files contain codes that describe their contents, such as 'AU' in place of 'author'. This function provides lookup tables for 'ris' tags of different kinds

tag_lookup(type = "ris")

Arguments

type

Which lookup table should be returned? Accepted values are 'ris', 'ris_write' or 'medline'

Value

a data.frame containing the original tag (column 'ris'), and the full-word substitution for that tag (column 'bib'). For type 'ris', there is also an added 'order' column showing the order those tags should be displayed in. 'ris_write' is a version of 'ris' with only one ris tag per bib tag.

Details

Primarily an internal function to support read_bibliography and write_bibliography. Tag substitutions for PubMed/Medline fields are taken directly from the NIH (available here). Substitutions for other ris-like formats are based on common examples, but are much less consistently documented.

See also

Examples

tag_lookup("ris") # standard ris format
#>    ris         bib order
#> 1   TY        type     1
#> 2   AU      author     2
#> 3   A1      author     2
#> 4   A2      author     2
#> 5   A3      author     2
#> 6   A4      author     2
#> 7   A5      author     2
#> 8   AF author_full     3
#> 9   PY        year     4
#> 10  Y1        year     4
#> 11  TI       title     5
#> 12  T1       title     5
#> 13  JO     journal     6
#> 14  T2     journal     6
#> 15  T3     journal     6
#> 16  SO     journal     6
#> 17  JT     journal     6
#> 18  JF     journal     6
#> 19  JA     journal     6
#> 20  VL      volume     7
#> 21  IS       issue     8
#> 22  EP       pages     9
#> 23  BP       pages     9
#> 24  SP       pages     9
#> 25  AB    abstract    10
#> 26  N2    abstract    10
#> 27  KW    keywords    11
#> 28  DE    keywords    11
#> 29  DO         doi    12
#> 30  DI         doi    12
#> 31  CN call_number    13
#> 32  SN        issn    14
#> 33  UR         url    15
#> 34  AN   accession    16
#> 35  CY institution    17
#> 36  PB   publisher    18
#> 37  PP    pubplace    19
#> 38  AD     address    20
#> 39  ED      editor    21
#> 40  ED     edition    22
#> 41  LA    language    23
#> 42  U1     eppi_id    24
tag_lookup("medline") # PubMed files
#>     ris                           bib
#> 1    AB                      abstract
#> 2    CI                copyright_info
#> 3    AD                   affiliation
#> 4  IRAD      investigator_affiliation
#> 5   AID                    article_id
#> 6    AU                        author
#> 7  AUID                     author_id
#> 8   FAU                   author_full
#> 9   BTI                    book_title
#> 10  CTI              collection_title
#> 11  COI          conflict_of_interest
#> 12   CN              author_corporate
#> 13 CRDT                  date_created
#> 14 DCOM                date_completed
#> 15   DA                  date_created
#> 16   LR                  date_revised
#> 17  DEP           date_published_elec
#> 18   DP                date_published
#> 19   EN                       edition
#> 20   ED                        editor
#> 21  FED                   editor_full
#> 22 EDAT                    date_added
#> 23   GS                   gene_symbol
#> 24   GN                  general_note
#> 25   GR                  grant_number
#> 26   IR                  investigator
#> 27  FIR             investigator_full
#> 28 ISBN                          isbn
#> 29   IS                          issn
#> 30   IP                         issue
#> 31   TA           journal_abbreviated
#> 32   JT                       journal
#> 33   LA                      language
#> 34  LID                   location_id
#> 35  MID                 manuscript_id
#> 36 MHDA                     mesh_date
#> 37   MH                    mesh_terms
#> 38  JID                        nlm_id
#> 39   RF                  references_n
#> 40  OAB                abstract_other
#> 41  OCI          copyright_info_other
#> 42  OID                      id_other
#> 43   OT                    term_other
#> 44  OTO              term_owner_other
#> 45  OWN                         owner
#> 46   PG                         pages
#> 47   PS      personal_name_as_subject
#> 48  FPS personal_name_as_subject_full
#> 49   PL               place_published
#> 50 PHST    publication_history_status
#> 51  PST            publication_status
#> 52   PT              publication_type
#> 53 PUBM              publishing_model
#> 54  PMC    pubmed_central_identitfier
#> 55 PMCR        pubmed_central_release
#> 56 PMID                     pubmed_id
#> 57   RN               registry_number
#> 58   NM                substance_name
#> 59   SI           secondary_source_id
#> 60   SO                        source
#> 61  SFM          space_flight_mission
#> 62 STAT                        status
#> 63   SB                        subset
#> 64   TI                         title
#> 65   TT          title_transliterated
#> 66   VI                        volume
#> 67  VTI                  volume_title